+++ pkg/auth/web/user.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/adrianpk/go-i18n/v2/i18n"
-	"gitlab.com/mikrowezel/backend/granica/internal/model"
+	"github.com/davecgh/go-spew/spew"
+	"github.com/nicksnyder/go-i18n/v2/i18n"
 	tp "gitlab.com/mikrowezel/backend/granica/pkg/auth/transport"
 	"gitlab.com/mikrowezel/backend/web"
 )
@@ -18,9 +18,19 @@ const (
 	UserCtxKey web.ContextKey = "user"
 )
 
+const (
+	createUserKey = "create-user"
+	updateUserKey = "update-user"
+)
+
 func (ep *Endpoint) InitCreateUser(w http.ResponseWriter, r *http.Request) {
+	// Retrieve stored form data if exists
+	// It avoids filling in the form again after submissions errors.
+	userFD := ep.restoreUserFD(r, createUserKey)
+
+	// Req & Res
 	res := &tp.CreateUserRes{}
-	res.FromModel(&model.User{}, "", nil)
+	res.FromTransport(&userFD, "", nil)
 	res.Action = ep.userCreateAction()
 
 	// Wrap response
@@ -29,15 +39,21 @@ func (ep *Endpoint) InitCreateUser(w http.ResponseWriter, r *http.Request) {
 	// Template
 	ts, err := ep.TemplateFor(userRes, web.CreateTmpl)
 	if err != nil {
-		ep.Redirect(w, r, "/")
+		m := ep.errMsg(r, web.CannotProcErr, userRes)
+		ep.storeErrorFlash(r, w, m)
+		ep.Redirect(w, r, UserPath())
+		ep.Log().Error(err)
 		return
 	}
 
 	// Write response
 	err = ts.Execute(w, wr)
 	if err != nil {
+		m := ep.errMsg(r, web.CannotProcErr, userRes)
+		ep.storeErrorFlash(r, w, m)
+		ep.Redirect(w, r, UserPath())
 		ep.Log().Error(err)
-		ep.Redirect(w, r, "/")
+		return
 	}
 }
 
@@ -47,52 +63,29 @@ func (ep *Endpoint) CreateUser(w http.ResponseWriter, r *http.Request) {
 
 	// TODO: Form data validation
 
+	// Store form data if exists
+	// It avoids filling in the form again after submissions errors.
+	ep.storeUserFD(r, w, createUserKey, req.User)
+
 	// Form to Req
 	err := web.FormToModel(r, &req.User)
-	res.Action = ep.userCreateAction()
-
-	// Template
-	ts, err := ep.TemplateFor(userRes, web.CreateTmpl)
 	if err != nil {
-		ep.Redirect(w, r, "/")
-		return
-	}
-
-	if err != nil {
-		m := ep.errMsg(r, web.CreatePfx, userRes)
-		wr := ep.ErrRes(r, res, m)
-		ep.Log().Error(err)
-
-		// TODO: Use redirect instead.
-		// Cleaner and avoids extra nesting level
-		// TODO: Preserve form data using gorilla session
-		err = ts.Execute(w, wr)
-		if err != nil {
-			ep.Log().Error(err)
-			ep.Redirect(w, r, "/")
-		}
+		ep.Redirect(w, r, UserPathNew())
 		return
 	}
 
 	// Service
 	err = ep.service.CreateUser(req, &res)
 	if err != nil {
-		wr := ep.ErrRes(r, res, "Cannot create user")
+		m := ep.errMsg(r, web.CreateErr, userRes)
+		ep.storeErrorFlash(r, w, m)
+		ep.Redirect(w, r, UserPathNew())
 		ep.Log().Error(err)
-
-		err = ts.Execute(w, wr)
-		if err != nil {
-			ep.Log().Error(err)
-			ep.Redirect(w, r, "/")
-		}
 		return
 	}
 
-	// TODO: Flash message
-
-	// Wrap response
+	ep.clearUserFD(r, w, createUserKey)
 	ep.Redirect(w, r, "/")
-
 }
 
 // GetUsers web endpoint.
@@ -149,9 +142,8 @@ func (ep *Endpoint) userUpdateAction(resource string, model web.Identifiable) we
 	return web.Action{Target: UserPathSlug(model), Method: "PUT"}
 }
 
-// NOTE: Just only testing some implementation path.
-// Not sure if this is the best one.
-func (ep *Endpoint) errMsg(r *http.Request, action, resource string) string {
+// Flash messages
+func (ep *Endpoint) errMsg(r *http.Request, errType, resource string) string {
 	l, ok := web.GetI18NLocalizer(r)
 	if !ok {
 		// FIX: Do something: Return default message?
@@ -159,13 +151,13 @@ func (ep *Endpoint) errMsg(r *http.Request, action, resource string) string {
 	}
 
 	// Message
-	id := fmt.Sprintf("%s_err_msg", action)
+	id := fmt.Sprintf("%s_err_msg", errType)
 
 	t, lang, err := l.LocalizeWithTag(&i18n.LocalizeConfig{
 		MessageID: id,
-		TemplateData: map[string]string{
-			"Name": resource,
-		},
+		//TemplateData: map[string]string{
+		//"Name": resource,
+		//},
 	})
 
 	if err != nil {
@@ -181,4 +173,45 @@ func (ep *Endpoint) localizeMessageID(l *i18n.Localizer, messageID string) (stri
 	return l.Localize(&i18n.LocalizeConfig{
 		MessageID: messageID,
 	})
+}
+
+// Flash
+func (ep *Endpoint) storeErrorFlash(r *http.Request, w http.ResponseWriter, message string) (ok bool) {
+	ep.Log().Debug("*Endpoint.storeErrorFlash not implemented")
+	return false
+}
+
+// Form data session helpers
+func (ep *Endpoint) storeUserFD(r *http.Request, w http.ResponseWriter, key string, userFD tp.User) (ok bool) {
+	s := ep.GetSession(r)
+	s.Values[key] = userFD
+	err := s.Save(r, w)
+	if err != nil {
+		return true
+	}
+	return false
+}
+
+func (ep *Endpoint) restoreUserFD(r *http.Request, key string) tp.User {
+	s := ep.GetSession(r)
+	v := s.Values[key]
+
+	user, ok := v.(tp.User)
+	if ok {
+		ep.Log().Debug("Stored form data", "value", spew.Sdump(user))
+		return user
+	}
+
+	ep.Log().Info("No stored form data", "key", key)
+	return tp.User{}
+}
+
+func (ep *Endpoint) clearUserFD(r *http.Request, w http.ResponseWriter, key string) (ok bool) {
+	s := ep.GetSession(r)
+	delete(s.Values, key)
+	err := s.Save(r, w)
+	if err != nil {
+		return true
+	}
+	return false
 }